anhnt cập nhập không cho xóa, ngừng hoạt động dung lượng nếu vẫn còn biến thể sản phẩm liên quan, button xóa đồng dạng 2 nút còn lại

+++ app/Http/Controllers/Admin/DungLuongController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Models\DungLuong;
 use Illuminate\Http\Request;
+use Illuminate\Validation\Rule;
 
 class DungLuongController extends Controller
 {
@@ -13,9 +14,9 @@ class DungLuongController extends Controller
      */
     public function index()
     {
-    
+
         $dungluongs = DungLuong::all();
-        return view('admins.dungluongs.index',compact('dungluongs'));
+        return view('admins.dungluongs.index', compact('dungluongs'));
     }
 
     /**
@@ -35,23 +36,25 @@ class DungLuongController extends Controller
         //
 
 
-        if($request->isMethod('POST')){
+        if ($request->isMethod('POST')) {
 
-            $validDungLuong = $request->validate([
-                'ten_dung_luong'=>'required|string|max:255,unique:dung_luongs,ten_dung_luong'
-            ],
-            [
-                'ten_dung_luong.required'=>'tên dung lượng không được để trống',
-                'ten_dung_luong.string'=>'tên dung lượng phải là một chuỗi!',
-                'ten_dung_luong.max'=>"Tên dung lượng không quá 255 ký tự!",
-                 'ten_dung_luong.unique'=>'Tên dung lượng đã tồn tại!'
-            ]);
+            $validDungLuong = $request->validate(
+                [
+                    'ten_dung_luong' => 'required|string|max:255|unique:dung_luongs,ten_dung_luong'
+                ],
+                [
+                    'ten_dung_luong.required' => 'tên dung lượng không được để trống',
+                    'ten_dung_luong.string' => 'tên dung lượng phải là một chuỗi!',
+                    'ten_dung_luong.max' => "Tên dung lượng không quá 255 ký tự!",
+                    'ten_dung_luong.unique' => 'Tên dung lượng đã tồn tại!'
+                ]
+            );
 
             $params = $request->except('_token');
 
             DungLuong::create($params);
 
-            return redirect()->route('admin.dungluongs.index')->with('success','Thêm mới dung lượng thành công');
+            return redirect()->route('admin.dungluongs.index')->with('success', 'Thêm mới dung lượng thành công');
         }
     }
 
@@ -71,7 +74,7 @@ class DungLuongController extends Controller
     {
         //
         $dungluongs = DungLuong::query()->findOrFail($id);
-        return view('admins.dungluongs.update',compact('dungluongs'));
+        return view('admins.dungluongs.update', compact('dungluongs'));
     }
 
     /**
@@ -80,26 +83,30 @@ class DungLuongController extends Controller
     public function update(Request $request, string $id)
     {
         //
-        if($request->isMethod("PUT")){
+        if ($request->isMethod("PUT")) {
 
             $validDungLuong = $request->validate([
-                'ten_dung_luong'=>'required|string|max:255|unique:dung_luongs,ten_dung_luong'
-            ],
-            [
-                'ten_dung_luong.required'=>'tên dung lượng không được để trống!',
-                'ten_dung_luong.string'=>'tên dung lượng phải là một chuỗi!',
-                'ten_dung_luong.max'=>"Tên dung lượng không quá 255 ký tự!",
-                'ten_dung_luong.unique'=>'Tên dung lượng đã tồn tại'
+                'ten_dung_luong' => [
+                    'required',
+                    'string',
+                    'max:255',
+                    Rule::unique('dung_luongs', 'ten_dung_luong')->ignore($id)
+                ]
+            ], [
+                'ten_dung_luong.required' => 'Tên dung lượng không được để trống!',
+                'ten_dung_luong.string' => 'Tên dung lượng phải là một chuỗi!',
+                'ten_dung_luong.max' => 'Tên dung lượng không quá 255 ký tự!',
+                'ten_dung_luong.unique' => 'Tên dung lượng đã tồn tại'
             ]);
 
 
-            $params = $request->except('_token','_method');
+            $params = $request->except('_token', '_method');
 
             $dungluongs = DungLuong::query()->findOrFail($id);
 
             $dungluongs->update($params);
 
-            return redirect()->route('admin.dungluongs.index')->with('success','cập nhật thành công');
+            return redirect()->route('admin.dungluongs.index')->with('success', 'cập nhật thành công');
         }
     }
 
@@ -110,6 +117,10 @@ class DungLuongController extends Controller
             return redirect()->route('admin.banners.index')->with('error', 'Dung lượng không tồn tại');
         }
         if ($dungluong->trang_thai == true) {
+            $countDungLuong = $dungluong->bienTheSanPhams()->withTrashed()->get();
+            if (count($countDungLuong) > 0) {
+                return redirect()->back()->with('error', 'Dung lượng đã có sản phẩm, không thể ngừng hoạt động!');
+            }
             $dungluong->trang_thai = false;
             $dungluong->save();
             return redirect()->back()->with('success', 'Ngừng hoạt động dung lượng');
@@ -127,9 +138,12 @@ class DungLuongController extends Controller
     {
         //
         $dungluongs = DungLuong::findOrFail($id);
-
+        $countDungLuong = $dungluongs->bienTheSanPhams()->withTrashed()->get();
+        if (count($countDungLuong) > 0) {
+            return redirect()->back()->with('error', 'Dung lượng đã có sản phẩm, không thể ngừng hoạt động!');
+        }
         $dungluongs->delete();
 
-        return redirect()->route('admin.dungluongs.index')->with('success','Xóa kích cỡ dung lượng thành công');
+        return redirect()->route('admin.dungluongs.index')->with('success', 'Xóa kích cỡ dung lượng thành công');
     }
 }
